Support latest DRDS Open API for version 2019-01-23.

+++ services/drds/struct_performance_value.go
@@ -18,5 +18,5 @@ package drds
 // PerformanceValue is a nested struct in drds response
 type PerformanceValue struct {
 	Value string `json:"Value" xml:"Value"`
-	Date  int    `json:"Date" xml:"Date"`
+	Date  int64  `json:"Date" xml:"Date"`
 }
